pyproject.toml
[project]
name = "stacked-datatable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pywebview>=6.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
]
